feat(ajouteClass): Removing the previous folder of activities
